requirements.txt
faiss_cpu==1.7.4
numpy==2.2.5
openai==1.78.0
pandas==2.2.3
python-dotenv==1.1.0
sentence_transformers==4.1.0
tqdm==4.67.1
